Crate voxrs_math

Structs

Aabb

Angle

BlockPos

ChunkPos

Frustum

IVec2
A 2-dimensional vector.

IVec3
A 3-dimensional vector.

Mat3
A 3x3 column major matrix.

Mat4
A 4x4 column major matrix.

Plane

Quat
A quaternion representing an orientation.

Ray

Rect2

Sphere

Vec2
A 2-dimensional vector.

Vec3
A 3-dimensional vector without SIMD support.

Vec4
A 4-dimensional vector.

WorldBlockCounts
block count in world in each direction (x, y, z)

WorldChunkCounts
chunk count in world in each direction (x, y, z)

Enums

Dir
XPos : X Positive direction , XNeg : X Negative direction

RayAabbResult
RayAabbResult Inside: ray is inside of aabb NotIntersect: not intersected Intersect: with distance and hit direction

Constants

BLOCK_COUNT_IN_CHUNKSIDE
block count in chunk direction (x, y, z)

TOTAL_BLOCK_COUNTS_IN_CHUNK
total block count in chunk

Functions

get_matrix
get matrix using row, and column index from 1 (row >= 1 && row <= 4, col >= 1 && col <= 4)

set_matrix
set matrix using row, and column index from 1 (row >= 1 && row <= 4, col >= 1 && col <= 4)