Struct voxrs_math::IVec3

#[repr(transparent)]
pub struct IVec3(_);

A 3-dimensional vector.

Implementations

impl IVec3

pub const ZERO: IVec3

All zeroes.

pub const ONE: IVec3

All ones.

pub const X: IVec3

[1, 0, 0]: a unit-length vector pointing along the positive X axis.

pub const Y: IVec3

[0, 1, 0]: a unit-length vector pointing along the positive Y axis.

pub const Z: IVec3

[0, 0, 1]: a unit-length vector pointing along the positive Z axis.

pub const AXES: [IVec3; 3]

The unit axes.

pub fn new(x: i32, y: i32, z: i32) -> IVec3

Creates a new 3D vector.

pub fn extend(self, w: i32) -> IVec4

Creates a 4D vector from self and the given w value.

pub fn truncate(self) -> IVec2

Creates a Vec2 from the x and y elements of self, discarding z.

Truncation may also be performed by using self.xy() or Vec2::from().

pub fn cross(self, other: IVec3) -> IVec3

Computes the cross product of self and other.

pub fn to_array(&self) -> [i32; 3]

[x, y, z]

pub fn splat(v: i32) -> IVec3

Creates a vector with all elements set to v.

pub fn select(mask: BVec3, if_true: IVec3, if_false: IVec3) -> IVec3

Creates a vector from the elements in if_true and if_false, selecting which to use for each element of self.

A true element in the mask uses the corresponding element from if_true, and false uses the element from if_false.

pub fn dot(self, other: IVec3) -> i32

Computes the dot product of self and other.

pub fn min(self, other: IVec3) -> IVec3

Returns a vector containing the mininum values for each element of self and other.

In other words this computes [self.x.max(other.x), self.y.max(other.y), ..].

pub fn max(self, other: IVec3) -> IVec3

Returns a vector containing the maximum values for each element of self and other.

In other words this computes [self.x.max(other.x), self.y.max(other.y), ..].

pub fn clamp(self, min: IVec3, max: IVec3) -> IVec3

Component-wise clamping of values, similar to f32::clamp.

Each element in min must be less-or-equal to the corresponing element in max.

If the glam-assert feature is enabled, the function will panic if the contract is not met, otherwise the behavior is undefined.

pub fn min_element(self) -> i32

Returns the horizontal minimum of self.

In other words this computes min(x, y, ..).

pub fn max_element(self) -> i32

Returns the horizontal maximum of self.

In other words this computes max(x, y, ..).

pub fn cmpeq(self, other: IVec3) -> BVec3

Returns a vector mask containing the result of a == comparison for each element of self and other.

In other words, this computes [self.x == other.x, self.y == other.y, ..] for all elements.

pub fn cmpne(self, other: IVec3) -> BVec3

Returns a vector mask containing the result of a != comparison for each element of self and other.

In other words this computes [self.x != other.x, self.y != other.y, ..] for all elements.

pub fn cmpge(self, other: IVec3) -> BVec3

Returns a vector mask containing the result of a >= comparison for each element of self and other.

In other words this computes [self.x >= other.x, self.y >= other.y, ..] for all elements.

pub fn cmpgt(self, other: IVec3) -> BVec3

Returns a vector mask containing the result of a > comparison for each element of self and other.

In other words this computes [self.x > other.x, self.y > other.y, ..] for all elements.

pub fn cmple(self, other: IVec3) -> BVec3

Returns a vector mask containing the result of a <= comparison for each element of self and other.

In other words this computes [self.x <= other.x, self.y <= other.y, ..] for all elements.

pub fn cmplt(self, other: IVec3) -> BVec3

Returns a vector mask containing the result of a < comparison for each element of self and other.

In other words this computes [self.x < other.x, self.y < other.y, ..] for all elements.

pub fn from_slice_unaligned(slice: &[i32]) -> IVec3

👎 Deprecated since 0.15.2:
Please use from_slice() instead

pub fn from_slice(slice: &[i32]) -> IVec3

Creates a vector from the first N values in slice.

Panics

Panics if slice is less than N elements long.

pub fn write_to_slice_unaligned(self, slice: &mut [i32])

👎 Deprecated since 0.15.2:
Please use write_to_slice() instead

pub fn write_to_slice(self, slice: &mut [i32])

Writes the elements of self to the first N elements in slice.

Panics

Panics if slice is less than N elements long.

pub fn abs(self) -> IVec3

Returns a vector containing the absolute value of each element of self.

pub fn signum(self) -> IVec3

Returns a vector with elements representing the sign of self.

1.0 if the number is positive, +0.0 or INFINITY

-1.0 if the number is negative, -0.0 or NEG_INFINITY

NAN if the number is NAN

pub fn as_f32(&self) -> Vec3

Casts all elements of self to f32.

pub fn as_f64(&self) -> DVec3

Casts all elements of self to f64.

pub fn as_u32(&self) -> UVec3

Casts all elements of self to u32.