Struct voxrs_math::Quat

#[repr(align(16))]
pub struct Quat(_);

A quaternion representing an orientation.

This quaternion is intended to be of unit length but may denormalize due to floating point “error creep” which can occur when successive quaternion operations are applied.

This type is 16 byte aligned.

Implementations

impl Quat

pub const IDENTITY: Quat

The identity quaternion. Corresponds to no rotation.

pub fn from_xyzw(x: f32, y: f32, z: f32, w: f32) -> Quat

Creates a new rotation quaternion.

This should generally not be called manually unless you know what you are doing. Use one of the other constructors instead such as identity or from_axis_angle.

from_xyzw is mostly used by unit tests and serde deserialization.

Preconditions

This function does not check if the input is normalized, it is up to the user to provide normalized input or to normalized the resulting quaternion.

pub fn from_array(a: [f32; 4]) -> Quat

Creates a rotation quaternion from an array.

Preconditions

This function does not check if the input is normalized, it is up to the user to provide normalized input or to normalized the resulting quaternion.

pub fn from_vec4(v: Vec4) -> Quat

Creates a new rotation quaternion from a 4D vector.

Preconditions

This function does not check if the input is normalized, it is up to the user to provide normalized input or to normalized the resulting quaternion.

pub fn from_slice_unaligned(slice: &[f32]) -> Quat

👎 Deprecated since 0.15.2:
Please use from_slice() instead

pub fn from_slice(slice: &[f32]) -> Quat

Creates a rotation quaternion from a slice.

Preconditions

This function does not check if the input is normalized, it is up to the user to provide normalized input or to normalized the resulting quaternion.

Panics

Panics if slice length is less than 4.

pub fn write_to_slice_unaligned(self, slice: &mut [f32])

👎 Deprecated since 0.15.2:
Please use write_to_slice() instead

pub fn write_to_slice(self, slice: &mut [f32])

Writes the quaternion to an unaligned slice.

Panics

Panics if slice length is less than 4.

pub fn from_axis_angle(axis: Vec3, angle: f32) -> Quat

Create a quaternion for a normalized rotation axis and angle (in radians). The axis must be normalized (unit-length).

pub fn from_scaled_axis(v: Vec3) -> Quat

Create a quaternion that rotates v.length() radians around v.normalize().

from_scaled_axis(Vec3::ZERO) results in the identity quaternion.

pub fn from_rotation_x(angle: f32) -> Quat

Creates a quaternion from the angle (in radians) around the x axis.

pub fn from_rotation_y(angle: f32) -> Quat

Creates a quaternion from the angle (in radians) around the y axis.

pub fn from_rotation_z(angle: f32) -> Quat

Creates a quaternion from the angle (in radians) around the z axis.

pub fn from_rotation_ypr(yaw: f32, pitch: f32, roll: f32) -> Quat

👎 Deprecated since 0.15.0:
Please use from_euler(EulerRot::YXZ, yaw, pitch, roll) instead

pub fn from_euler(euler: EulerRot, a: f32, b: f32, c: f32) -> Quat

Creates a quaternion from the given euler rotation sequence and the angles (in radians).

pub fn from_rotation_mat3(mat: &Mat3) -> Quat

👎 Deprecated since 0.15.0:
Please use from_mat3 instead

pub fn from_mat3(mat: &Mat3) -> Quat

Creates a quaternion from a 3x3 rotation matrix.

pub fn from_rotation_mat4(mat: &Mat4) -> Quat

👎 Deprecated since 0.15.0:
Please use from_mat4 instead

pub fn from_mat4(mat: &Mat4) -> Quat

Creates a quaternion from a 3x3 rotation matrix inside a homogeneous 4x4 matrix.

pub fn from_rotation_arc(from: Vec3, to: Vec3) -> Quat

Gets the minimal rotation for transforming from to to. The rotation is in the plane spanned by the two vectors. Will rotate at most 180 degrees.

The input vectors must be normalized (unit-length).

from_rotation_arc(from, to) * from ≈ to.

For near-singular cases (from≈to and from≈-to) the current implementation is only accurate to about 0.001 (for f32).

pub fn from_rotation_arc_colinear(from: Vec3, to: Vec3) -> Quat

Gets the minimal rotation for transforming from to either to or -to. This means that the resulting quaternion will rotate from so that it is colinear with to.

The rotation is in the plane spanned by the two vectors. Will rotate at most 90 degrees.

The input vectors must be normalized (unit-length).

to.dot(from_rotation_arc_colinear(from, to) * from).abs() ≈ 1.

pub fn to_axis_angle(self) -> (Vec3, f32)

Returns the rotation axis and angle (in radians) of self.

pub fn to_scaled_axis(self) -> Vec3

Returns the rotation axis scaled by the rotation in radians.

pub fn to_euler(self, euler: EulerRot) -> (f32, f32, f32)

Returns the rotation angles for the given euler rotation sequence.

#[must_use]
pub fn conjugate(self) -> Quat

Returns the quaternion conjugate of self. For a unit quaternion the conjugate is also the inverse.

#[must_use]
pub fn inverse(self) -> Quat

Returns the inverse of a normalized quaternion.

Typically quaternion inverse returns the conjugate of a normalized quaternion. Because self is assumed to already be unit length this method does not normalize before returning the conjugate.

pub fn dot(self, other: Quat) -> f32

Computes the dot product of self and other. The dot product is equal to the the cosine of the angle between two quaternion rotations.

pub fn length(self) -> f32

Computes the length of self.

pub fn length_squared(self) -> f32

Computes the squared length of self.

This is generally faster than length() as it avoids a square root operation.

pub fn length_recip(self) -> f32

Computes 1.0 / length().

For valid results, self must not be of length zero.

#[must_use]
pub fn normalize(self) -> Quat

Returns self normalized to length 1.0.

For valid results, self must not be of length zero.

pub fn is_finite(self) -> bool

Returns true if, and only if, all elements are finite. If any element is either NaN, positive or negative infinity, this will return false.

pub fn is_nan(self) -> bool

pub fn is_normalized(self) -> bool

Returns whether self of length 1.0 or not.

Uses a precision threshold of 1e-6.

pub fn is_near_identity(self) -> bool

pub fn angle_between(self, other: Quat) -> f32

Returns the angle (in radians) for the minimal rotation for transforming this quaternion into another.

Both quaternions must be normalized.

pub fn abs_diff_eq(self, other: Quat, max_abs_diff: f32) -> bool

Returns true if the absolute difference of all elements between self and other is less than or equal to max_abs_diff.

This can be used to compare if two quaternions contain similar elements. It works best when comparing with a known value. The max_abs_diff that should be used used depends on the values being compared against.

For more see comparing floating point numbers.

pub fn lerp(self, end: Quat, s: f32) -> Quat

Performs a linear interpolation between self and other based on the value s.

When s is 0.0, the result will be equal to self. When s is 1.0, the result will be equal to other.

pub fn slerp(self, end: Quat, s: f32) -> Quat

Performs a spherical linear interpolation between self and end based on the value s.

When s is 0.0, the result will be equal to self. When s is 1.0, the result will be equal to end.

Note that a rotation can be represented by two quaternions: q and -q. The slerp path between q and end will be different from the path between -q and end. One path will take the long way around and one will take the short way. In order to correct for this, the dot product between self and end should be positive. If the dot product is negative, slerp between -self and end.

pub fn mul_vec3(self, other: Vec3) -> Vec3

Multiplies a quaternion and a 3D vector, returning the rotated vector.

pub fn mul_quat(self, other: Quat) -> Quat

Multiplies two quaternions. If they each represent a rotation, the result will represent the combined rotation. Note that due to floating point rounding the result may not be perfectly normalized.

pub fn mul_vec3a(self, other: Vec3A) -> Vec3A

Multiplies a quaternion and a 3D vector, returning the rotated vector.

pub fn as_f64(self) -> DQuat

pub fn from_affine3(mat: &Affine3A) -> Quat

Creates a quaternion from a 3x3 rotation matrix inside a 3D affine transform.