Struct voxrs_math::Vec2

#[repr(transparent)]
pub struct Vec2(_);

A 2-dimensional vector.

Implementations

impl Vec2

pub const ZERO: Vec2

All zeroes.

pub const ONE: Vec2

All ones.

pub const X: Vec2

[1, 0]: a unit-length vector pointing along the positive X axis.

pub const Y: Vec2

[0, 1]: a unit-length vector pointing along the positive Y axis.

pub const AXES: [Vec2; 2]

The unit axes.

pub fn new(x: f32, y: f32) -> Vec2

Creates a new vector.

pub fn extend(self, z: f32) -> Vec3

Creates a 3D vector from self and the given z value.

pub fn to_array(&self) -> [f32; 2]

[x, y]

pub fn splat(v: f32) -> Vec2

Creates a vector with all elements set to v.

pub fn select(mask: BVec2, if_true: Vec2, if_false: Vec2) -> Vec2

Creates a vector from the elements in if_true and if_false, selecting which to use for each element of self.

A true element in the mask uses the corresponding element from if_true, and false uses the element from if_false.

pub fn dot(self, other: Vec2) -> f32

Computes the dot product of self and other.

pub fn min(self, other: Vec2) -> Vec2

Returns a vector containing the mininum values for each element of self and other.

In other words this computes [self.x.max(other.x), self.y.max(other.y), ..].

pub fn max(self, other: Vec2) -> Vec2

Returns a vector containing the maximum values for each element of self and other.

In other words this computes [self.x.max(other.x), self.y.max(other.y), ..].

pub fn clamp(self, min: Vec2, max: Vec2) -> Vec2

Component-wise clamping of values, similar to f32::clamp.

Each element in min must be less-or-equal to the corresponing element in max.

If the glam-assert feature is enabled, the function will panic if the contract is not met, otherwise the behavior is undefined.

pub fn min_element(self) -> f32

Returns the horizontal minimum of self.

In other words this computes min(x, y, ..).

pub fn max_element(self) -> f32

Returns the horizontal maximum of self.

In other words this computes max(x, y, ..).

pub fn cmpeq(self, other: Vec2) -> BVec2

Returns a vector mask containing the result of a == comparison for each element of self and other.

In other words, this computes [self.x == other.x, self.y == other.y, ..] for all elements.

pub fn cmpne(self, other: Vec2) -> BVec2

Returns a vector mask containing the result of a != comparison for each element of self and other.

In other words this computes [self.x != other.x, self.y != other.y, ..] for all elements.

pub fn cmpge(self, other: Vec2) -> BVec2

Returns a vector mask containing the result of a >= comparison for each element of self and other.

In other words this computes [self.x >= other.x, self.y >= other.y, ..] for all elements.

pub fn cmpgt(self, other: Vec2) -> BVec2

Returns a vector mask containing the result of a > comparison for each element of self and other.

In other words this computes [self.x > other.x, self.y > other.y, ..] for all elements.

pub fn cmple(self, other: Vec2) -> BVec2

Returns a vector mask containing the result of a <= comparison for each element of self and other.

In other words this computes [self.x <= other.x, self.y <= other.y, ..] for all elements.

pub fn cmplt(self, other: Vec2) -> BVec2

Returns a vector mask containing the result of a < comparison for each element of self and other.

In other words this computes [self.x < other.x, self.y < other.y, ..] for all elements.

pub fn from_slice_unaligned(slice: &[f32]) -> Vec2

ðŸ‘Ž Deprecated since 0.15.2:
Please use from_slice() instead

pub fn from_slice(slice: &[f32]) -> Vec2

Creates a vector from the first N values in slice.

Panics

Panics if slice is less than N elements long.

pub fn write_to_slice_unaligned(self, slice: &mut [f32])

ðŸ‘Ž Deprecated since 0.15.2:
Please use write_to_slice() instead

pub fn write_to_slice(self, slice: &mut [f32])

Writes the elements of self to the first N elements in slice.

Panics

Panics if slice is less than N elements long.

pub fn abs(self) -> Vec2

Returns a vector containing the absolute value of each element of self.

pub fn signum(self) -> Vec2

Returns a vector with elements representing the sign of self.

1.0 if the number is positive, +0.0 or INFINITY

-1.0 if the number is negative, -0.0 or NEG_INFINITY

NAN if the number is NAN

pub fn perp(self) -> Vec2

Returns a vector that is equal to self rotated by 90 degrees.

pub fn perp_dot(self, other: Vec2) -> f32

The perpendicular dot product of self and other. Also known as the wedge product, 2d cross product, and determinant.

pub fn is_finite(self) -> bool

Returns true if, and only if, all elements are finite. If any element is either NaN, positive or negative infinity, this will return false.

pub fn is_nan(self) -> bool

Returns true if any elements are NaN.

pub fn is_nan_mask(self) -> BVec2

Performs is_nan on each element of self, returning a vector mask of the results.

In other words, this computes [x.is_nan(), y.is_nan(), z.is_nan(), w.is_nan()].

pub fn length(self) -> f32

Computes the length of self.

pub fn length_squared(self) -> f32

Computes the squared length of self.

This is faster than length() as it avoids a square root operation.

pub fn length_recip(self) -> f32

Computes 1.0 / length().

For valid results, self must not be of length zero.

pub fn distance(self, other: Vec2) -> f32

Computes the Euclidean distance between two points in space.

pub fn distance_squared(self, other: Vec2) -> f32

Compute the squared euclidean distance between two points in space.

#[must_use]
pub fn normalize(self) -> Vec2

Returns self normalized to length 1.0.

For valid results, self must not be of length zero, nor very close to zero.

See also Self::try_normalize and Self::normalize_or_zero.

#[must_use]
pub fn try_normalize(self) -> Option<Vec2>

Returns self normalized to length 1.0 if possible, else returns None.

In particular, if the input is zero (or very close to zero), or non-finite, the result of this operation will be None.

See also Self::normalize_or_zero.

#[must_use]
pub fn normalize_or_zero(self) -> Vec2

Returns self normalized to length 1.0 if possible, else returns zero.

In particular, if the input is zero (or very close to zero), or non-finite, the result of this operation will be zero.

See also Self::try_normalize.

pub fn is_normalized(self) -> bool

Returns whether self is length 1.0 or not.

Uses a precision threshold of 1e-6.

pub fn round(self) -> Vec2

Returns a vector containing the nearest integer to a number for each element of self. Round half-way cases away from 0.0.

pub fn floor(self) -> Vec2

Returns a vector containing the largest integer less than or equal to a number for each element of self.

pub fn ceil(self) -> Vec2

Returns a vector containing the smallest integer greater than or equal to a number for each element of self.

pub fn fract(self) -> Vec2

Returns a vector containing the fractional part of the vector, e.g. self - self.floor().

Note that this is fast but not precise for large numbers.

pub fn exp(self) -> Vec2

Returns a vector containing e^self (the exponential function) for each element of self.

pub fn powf(self, n: f32) -> Vec2

Returns a vector containing each element of self raised to the power of n.

pub fn recip(self) -> Vec2

Returns a vector containing the reciprocal 1.0/n of each element of self.

pub fn lerp(self, other: Vec2, s: f32) -> Vec2

Performs a linear interpolation between self and other based on the value s.

When s is 0.0, the result will be equal to self. When s is 1.0, the result will be equal to other.

pub fn abs_diff_eq(self, other: Vec2, max_abs_diff: f32) -> bool

Returns true if the absolute difference of all elements between self and other is less than or equal to max_abs_diff.

This can be used to compare if two vectors contain similar elements. It works best when comparing with a known value. The max_abs_diff that should be used used depends on the values being compared against.

For more see comparing floating point numbers.

pub fn clamp_length(self, min: f32, max: f32) -> Vec2

Returns a vector with a length no less than min and no more than max

pub fn clamp_length_max(self, max: f32) -> Vec2

Returns a vector with a length no more than max

pub fn clamp_length_min(self, min: f32) -> Vec2

Returns a vector with a length no less than min

pub fn angle_between(self, other: Vec2) -> f32

Returns the angle (in radians) between self and other.

The input vectors do not need to be unit length however they must be non-zero.

pub fn as_f64(&self) -> DVec2

Casts all elements of self to f64.

pub fn as_i32(&self) -> IVec2

Casts all elements of self to i32.

pub fn as_u32(&self) -> UVec2

Casts all elements of self to u32.